+++ Back-end/history.txt
@@ -1,6 +1,3 @@
 User Input >> 사용자가 앱을 실행했습니다.
-AI Response >>  안녕하세요. 리쿠와 함께하는 키오스크 교육에 오신 것을 환영합니다. 오늘 해볼 미션은 [당수동]로 가는 [18:30] 버스를 선택하고 [어른 1명, 중고생 1명, 아동 1명]을 예매하는 것입니다. 화면 오른쪽 보라색 버튼을 눌러 목적지 선택 화면으로 이동해볼까요?
-
-User Input >> 사용자는 처음 화면에서 보라색 버튼을 눌렀습니다. 목적지 화면으로 이동합니다.
-AI Response >>  상단의 "경기도"를 선택하세요. 그리고 초성 "ㄷ"을 선택하고 마지막으로 "당수동"을 선택하세요.
+AI Response >>  안녕하세요. 리쿠와 함께하는 키오스크 교육에 오신 것을 환영합니다. 오늘 해볼 미션은 [부산해운대(수도권)] 지역으로 가는 [18:30] 버스를 선택하고 [어른 2명, 아동 1명]을 예매하는 것입니다. 화면 오른쪽 보라색 버튼을 눌러 목적지 선택 화면으로 이동해볼까요?
