+++ Back-end/main.py
@@ -7,7 +7,64 @@ from pathlib import Path  # 파일 경로 확인을 위한 Path 사용
 from typing import List, Dict
 import json
 import random
+from dotenv import load_dotenv
+import os
+import paho.mqtt.client as mqtt
 import re
+
+load_dotenv()
+openai_api_key = os.getenv("OPENAI_API_KEY")
+
+broker_ip = "20.41.118.19"
+port = 1883
+topic = "liku/809b335b-4c49-4a92-8db2-5d79e4b7a890/json/list"  # Publish할 topic
+
+
+def publish_mqtt_message(response: str):
+    client = mqtt.Client()  # MQTT 클라이언트 생성
+    
+    # 콜백 설정
+    client.on_connect = lambda client, userdata, flags, rc: print("MQTT 연결 성공" if rc == 0 else f"MQTT 연결 실패, 코드: {rc}")
+    client.on_publish = lambda client, userdata, mid: print("메시지 발행 성공")
+
+    # MQTT 서버 연결
+    client.connect(broker_ip, port, 60)
+
+    # JSON payload 생성
+    payload = {
+        "data": [
+            {
+                "action": {
+                    "action_name": "fighting",
+                    "action_type": "overlap",
+                    "pitch": 0,
+                    "yaw": 0
+                },
+                "display": {
+                    "display_name": "blink",
+                    "delay": 0,
+                    "playtime": 1,
+                    "playspeed": 3
+                },
+                "speech": {
+                    "speech_name": "TTS_output",
+                    "TTS": response,  # 리쿠의 답변을 TTS에 넣음
+                    "delay": 0,
+                    "repeat": 0
+                },
+                "listen": False
+            }
+        ]
+    }
+    
+    # MQTT로 payload 발행
+    payload_str = json.dumps(payload, ensure_ascii=False)
+    client.publish(topic, payload_str)
+    print('메시지 발행 완료 >> ', payload_str)
+
+    # 연결 종료
+    client.disconnect()
+
 
 # OpenAI 클라이언트 및 FastAPI 설정
 app = FastAPI()
@@ -28,8 +85,8 @@ class Message(BaseModel):
 # gpt-4o-mini
 # gpt-4o
 # LangChain 설정
-llm = ChatOpenAI(model= 'gpt-4o', temperature=0.5)
-llm_mini = ChatOpenAI(model='gpt-4o-mini', temperature=0.7)
+llm = ChatOpenAI(api_key=openai_api_key, model= 'gpt-4o', temperature=0.3)
+llm_mini = ChatOpenAI(api_key=openai_api_key, model='gpt-4o-mini', temperature=0.7)
 
 # 파일 경로 확인
 json_file_path = Path('kiosk.json')
@@ -66,7 +123,7 @@ def generate_scenario():
 @app.on_event("startup")
 def create_scenario_on_startup():
     scenario = generate_scenario()
-    print("\n서버가 실행되었습니다.\n")
+    print("\n 서버가 실행되었습니다. \n")
     global scenario_content
     with open('scenario.txt', 'r', encoding='utf-8') as file:
         scenario_content = file.read()
@@ -110,16 +167,16 @@ Kiosk Context:
 User Manual:
 {manual}
 
-The scenario that the user must follow:
+The scenario that the user must follow::
 {scenario}
 
-User input Log:
+User input & AI response Log
 {history}
 
 Button information pressed by the user:
 {question}
 
-답변 (한국어로, 1줄로):
+답변 (한국어로, 1줄만):
 """
 prompt = PromptTemplate(template=prompt_template, input_variables=["context", "manual", "scenario", "history", "question"])
 
@@ -131,8 +188,9 @@ def ask_question(query):
 
     # 프롬프트에 질문과 문서 내용 연결
     answer = qa_chain.invoke({"context": context,  "manual": manual_content, "scenario": scenario_content, "history": history_content, "question": query})
-
+    
     return answer.content
+
 
 with open('history.txt', 'w', encoding='utf-8') as file:
     pass
@@ -143,7 +201,7 @@ def clean_history():
 
 def compliment_generate(response):
     # 칭찬 생성 프롬프트 설정
-    compliment_prompt = PromptTemplate(template="{response}\n위의 응답내용의 앞이나 뒤에 응답 내용과 어울리는 칭찬이나 격려를 한마디만 더해주세요. 시나리오와 다른 선택을 했다면 아쉽습니다. 이런멘트로 짧게 해줘. 단, 응답 내용을 전부 그대로 적어줘야 합니다. Example. 괜찮아요. 다시한번 생각해볼까요? 잘하고 있습니다! etc",
+    compliment_prompt = PromptTemplate(template="{response}\n위의 응답내용의 앞이나 뒤에 응답 내용과 어울리는 칭찬이나 격려를 한마디만 더해주세요. 단, 응답 내용을 전부 그대로 적어줘야 합니다. Example. 괜찮아요. 다시한번 생각해볼까요? 잘하고 있습니다! etc",
     input_variables=["response"]
     ).format(response=response)
 
@@ -152,8 +210,6 @@ def compliment_generate(response):
 
     # Return the content generated by gpt-4o-mini
     return final_response.content
-
-
 
 # FastAPI 라우트 정의
 @app.post("/chat")
@@ -168,15 +224,15 @@ async def chat(message: Message):
         bus_time = re.search(r'버스시간 = "(.*?)"', scenario_content).group(1)
         people = re.search(r'인원 = "(.*?)"', scenario_content).group(1)
         # Format the response
-        response = f"안녕하세요. 리쿠와 함께하는 키오스크 교육에 오신 것을 환영합니다. 오늘 해볼 미션은 [{destination}]로 가는 [{bus_time}] 버스를 선택하고 [{people}]을 예매하는 것입니다. 화면 오른쪽 보라색 버튼을 눌러 목적지 선택 화면으로 이동해볼까요?"
+        response = f"안녕하세요. 리쿠와 함께하는 키오스크 교육에 오신 것을 환영합니다. 오늘 해볼 미션은 [{destination}] 지역으로 가는 [{bus_time}] 버스를 선택하고 [{people}]을 예매하는 것입니다. 화면 오른쪽 보라색 버튼을 눌러 목적지 선택 화면으로 이동해볼까요?"
         clean_history()  # Clear history if needed
         print('사용자 입력 >> ' + message.content)
         print('AI 답변 >> ' + response)
         with open('history.txt', 'a', encoding='utf-8') as file:
             file.write(f'User Input >> {message.content}\n')
             file.write(f'AI Response >>  {response}\n\n')
-        return response
-        
+        publish_mqtt_message(response)
+        return response        
     if message.content == "RESTART":
         scenario = generate_scenario()
         print(f"생성된 시나리오:\n{scenario}")
@@ -185,6 +241,7 @@ async def chat(message: Message):
             scenario_content = file.read()
         response = ask_question(message.content)
         print(response)
+        publish_mqtt_message(response)
         clean_history()
         return response
     
@@ -199,15 +256,13 @@ async def chat(message: Message):
         file.write(f'User Input >> {message.content}\n')
         file.write(f'AI Response >>  {response}\n\n')
     # 답변을 출력
-
+    publish_mqtt_message(response)
     return response
 
-# 시나리오 텍스트 확인용 라우트
 @app.get("/scenario")
 async def get_scenario():
     with open('scenario.txt', 'r', encoding='utf-8') as file:
         scenario = file.read()
     return {"scenario": scenario}
-    
 # uvicorn main:app --reload
 # uvicorn main:app --reload --host 0.0.0.0 --port 8000
