+++ Back-end/scenario.txt
@@ -1,3 +1,3 @@
-목적지 = "당수동"
+목적지 = "부산해운대(수도권)"
 버스시간 = "18:30"
-인원 = "어른 1명, 중고생 1명, 아동 1명"
+인원 = "어른 2명, 아동 1명"
